% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncapi-package.r
\docType{data}
\name{nc_data_types}
\alias{nc_data_types}
\title{NetCDF types}
\description{
A data frame of the names of the NetCDF data types. Column `name` is the
descriptive name, and `id` is the 0-based index, corresponding to the value returned by
nc_inq_var.
}
\details{
See code in data-raw/ for the creation.

http://www.unidata.ucar.edu/software/netcdf/docs/data_type.html
}
