% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_types.R
\name{nc_types}
\alias{nc_types}
\title{NetCDF types}
\usage{
nc_types(x)
}
\arguments{
\item{x}{id integer, or character coercible to integer}
}
\value{
type names
}
\description{
Return the descriptive name of the NetCDF data type from its ID number.
The ID is 0-based.
http://www.unidata.ucar.edu/software/netcdf/docs/data_type.html
}
