% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Rnc_close}
\alias{Rnc_close}
\title{Close a connection.}
\usage{
Rnc_close(ncid)
}
\arguments{
\item{ncid}{file connection provided by `Rnc_open`}
}
\description{
Close a connection.
}
\examples{
f_l3b <- system.file("extdata", "oceandata", "S2008001.L3b_DAY_CHL.nc", package = "ncapi")
 con <- Rnc_open(f_l3b)
 Rnc_close(con)
}
