% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Rnc_inq}
\alias{Rnc_inq}
\title{Source inquiry}
\usage{
Rnc_inq(grpid)
}
\arguments{
\item{grpid}{group ID provided by `Rnc_inq_grps`}
}
\description{
once we have a given ID (group-less file, or specific group)
find its contents
}
\examples{
f_l3b <- system.file("extdata", "oceandata", "S2008001.L3b_DAY_CHL.nc", package = "ncapi")
 con <- Rnc_open(f_l3b)
 groupids <- Rnc_inq_grps(con)
 l3b <- Rnc_inq(groupids[1])
 Rnc_close(con)
 print(f_l3b)
 print(l3b)
}
