% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Rnc_inq_dimension}
\alias{Rnc_inq_dimension}
\title{Dimension inquiry}
\usage{
Rnc_inq_dimension(grpid)
}
\arguments{
\item{grpid}{group ID provided by `Rnc_inq_grps`}
}
\description{
Dimension inquiry
}
\examples{
f_l3b <- system.file("extdata", "oceandata", "S2008001.L3b_DAY_CHL.nc", package = "ncapi")
con <- Rnc_open(f_l3b)
groupids <- Rnc_inq_grps(con)
Rnc_inq_dimension(groupids[1])
Rnc_close(con)
f_l3m <- system.file("extdata", "oceandata", "S2008001.L3m_DAY_CHL_chlor_a_9km.nc", package = "ncapi")

}
