% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Rnc_inq_dims}
\alias{Rnc_inq_dims}
\title{Dimension inquiry}
\usage{
Rnc_inq_dims(grpid)
}
\arguments{
\item{grpid}{group ID provided by `Rnc_inq_grps`}
}
\description{
Dimension inquiry
}
\examples{
f_l3b <- system.file("extdata", "oceandata", "S2008001.L3b_DAY_CHL.nc", package = "ncapi")
con <- Rnc_open(f_l3b)
groupids <- Rnc_inq_grps(con)
Rnc_inq_dims(groupids[1])
Rnc_close(con)
f_l3m <- system.file("extdata", "oceandata", "S2008001.L3m_DAY_CHL_chlor_a_9km.nc", package = "ncapi")
# ncmeta::nc_dims(f_l3m)
# A tibble: 4 x 4
      id          name length unlim
 <int>         <chr>  <dbl> <lgl>
 1     0           lat   2160 FALSE
 2     1           lon   4320 FALSE
 3     2           rgb      3 FALSE
 4     3 eightbitcolor    256 FALSE
}
