% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Rnc_inq_grpname}
\alias{Rnc_inq_grpname}
\title{Inquire group name by ID}
\usage{
Rnc_inq_grpname(grpid)
}
\arguments{
\item{grpid}{group ID provided by `Rnc_inq_grps`}
}
\description{
Inquire group name by ID
}
\examples{
f_l3b <- system.file("extdata", "oceandata", "S2008001.L3b_DAY_CHL.nc", package = "ncapi")
 con <- Rnc_open(f_l3b)
 lapply(Rnc_inq_grps(con), Rnc_inq_grpname)
 Rnc_close(con)
}
