% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Rnc_inq_natts}
\alias{Rnc_inq_natts}
\title{We already know this from Rnc_inq_variable}
\usage{
Rnc_inq_natts(grpid, varid)
}
\arguments{
\item{grpid}{con}

\item{varid}{variable id}
}
\description{
We already know this from Rnc_inq_variable
}
