% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Rnc_inq_vardims}
\alias{Rnc_inq_vardims}
\title{Dimensions of variables}
\usage{
Rnc_inq_vardims(grpid, ivar)
}
\arguments{
\item{grpid}{group ID provided by `Rnc_inq_grps`}

\item{ivar}{index of variable}
}
\description{
Dimensions of variables
}
\examples{
f_l3m <- system.file("extdata", "oceandata", "S2008001.L3m_DAY_CHL_chlor_a_9km.nc", package = "ncapi")
con <- Rnc_open(f_l3m)
variables <- tibble::as_tibble(Rnc_inq_variable(con))
 setNames(lapply(variables$id, function(x) Rnc_inq_vardims(con, x)), variables$name)
Rnc_close(con)
}
