% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Rnc_inq_variable}
\alias{Rnc_inq_variable}
\title{Variable inquiry}
\usage{
Rnc_inq_variable(grpid)
}
\arguments{
\item{grpid}{group ID provided by `Rnc_inq_grps`}
}
\description{
Variable inquiry
}
\examples{
f_l3m <- system.file("extdata", "oceandata", "S2008001.L3m_DAY_CHL_chlor_a_9km.nc", package = "ncapi")
con <- Rnc_open(f_l3m)
tibble::as_tibble(Rnc_inq_variable(con))
Rnc_close(con)
## that should be the same as
# ncmeta::nc_vars(f_l3m)
# # A tibble: 4 x 5
#   id    name  type ndims natts
#   <int>   <chr> <chr> <int> <int>
#   0     chlor_a float     2    12
#   1     lat float     1     5
#   2     lon float     1     5
#   3     palette ubyte     2     0
}
