% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Rnc_open}
\alias{Rnc_open}
\title{Open a connection.}
\usage{
Rnc_open(dsn)
}
\arguments{
\item{dsn}{data source name, i.e. file path or server path}
}
\description{
Pass  a character data source name to the internal nc_open library function.
}
\examples{
f_l3b <- system.file("inst", "oceandata", "S2008001.L3b_DAY_CHL.nc", package = "ncapi")
 Rnc_open(f_l3b)
f_l3m <- system.file("inst", "oceandata", "S2008001.L3m_DAY_CHL_chlor_a_9km.nc", package = "ncapi")
 Rnc_open(f_l3m)
u_cst <- "http://coastwatch.pfeg.noaa.gov/erddap/griddap/erdQSwind3da"
 Rnc_open(u_cst)
}
