% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Rnc_inq_att}
\alias{Rnc_inq_att}
\title{Attribute details}
\usage{
Rnc_inq_att(grpid, varid, attid)
}
\arguments{
\item{grpid}{con}

\item{varid}{variable id (can be global)}

\item{attid}{attribute id (within variable)}
}
\description{
Attribute details
}
